executor: move linux specific universal executor funcs to seperate file
