+++ client/driver/executor/executor_linux.go
@@ -20,7 +20,6 @@ import (
 	dstructs "github.com/hashicorp/nomad/client/driver/structs"
 	"github.com/hashicorp/nomad/client/stats"
 	cstructs "github.com/hashicorp/nomad/client/structs"
-	"github.com/hashicorp/nomad/helper"
 	"github.com/hashicorp/nomad/helper/discover"
 	shelpers "github.com/hashicorp/nomad/helper/stats"
 	"github.com/hashicorp/nomad/helper/uuid"
@@ -279,24 +278,28 @@ func (l *LibcontainerExecutor) Shutdown(signal string, grace time.Duration) erro
 		return nil
 	}
 
-	if signal == "" {
-		signal = "SIGINT"
-	}
+	if grace > 0 {
+		if signal == "" {
+			signal = "SIGINT"
+		}
 
-	sig, ok := signals.SignalLookup[signal]
-	if !ok {
-		return fmt.Errorf("error unknown signal given for shutdown: %s", signal)
-	}
+		sig, ok := signals.SignalLookup[signal]
+		if !ok {
+			return fmt.Errorf("error unknown signal given for shutdown: %s", signal)
+		}
 
-	err = l.container.Signal(sig, false)
-	if err != nil {
-		return err
-	}
+		err = l.container.Signal(sig, false)
+		if err != nil {
+			return err
+		}
 
-	select {
-	case <-l.userProcExited:
-		return nil
-	case <-time.After(grace):
+		select {
+		case <-l.userProcExited:
+			return nil
+		case <-time.After(grace):
+			return l.container.Signal(os.Kill, false)
+		}
+	} else {
 		return l.container.Signal(os.Kill, false)
 	}
 }
@@ -578,21 +581,6 @@ func newLibcontainerConfig(command *ExecCommand) *lconfigs.Config {
 	return cfg
 }
 
-// configureResourceContainer
-func (e *UniversalExecutor) configureResourceContainer(pid int) error {
-	cfg := &lconfigs.Config{
-		Cgroups: &lconfigs.Cgroup{
-			Resources: &lconfigs.Resources{
-				AllowAllDevices: helper.BoolToPtr(true),
-			},
-		},
-	}
-
-	configureBasicCgroups(cfg)
-	e.resConCtx.groups = cfg.Cgroups
-	return cgroups.EnterPid(cfg.Cgroups.Paths, pid)
-}
-
 // JoinRootCgroup moves the current process to the cgroups of the init process
 func JoinRootCgroup(subsystems []string) error {
 	mErrs := new(multierror.Error)
@@ -612,80 +600,5 @@ func JoinRootCgroup(subsystems []string) error {
 		multierror.Append(mErrs, err)
 	}
 
-	return mErrs.ErrorOrNil()
-}
-
-// destroyCgroup kills all processes in the cgroup and removes the cgroup
-// configuration from the host. This function is idempotent.
-func DestroyCgroup(groups *lconfigs.Cgroup, executorPid int) error {
-	mErrs := new(multierror.Error)
-	if groups == nil {
-		return fmt.Errorf("Can't destroy: cgroup configuration empty")
-	}
-
-	// Move the executor into the global cgroup so that the task specific
-	// cgroup can be destroyed.
-	path, err := cgroups.GetInitCgroupPath("freezer")
-	if err != nil {
-		return err
-	}
-
-	if err := cgroups.EnterPid(map[string]string{"freezer": path}, executorPid); err != nil {
-		return err
-	}
-
-	// Freeze the Cgroup so that it can not continue to fork/exec.
-	groups.Resources.Freezer = lconfigs.Frozen
-	freezer := cgroupFs.FreezerGroup{}
-	if err := freezer.Set(groups.Paths[freezer.Name()], groups); err != nil {
-		return err
-	}
-
-	var procs []*os.Process
-	pids, err := cgroups.GetAllPids(groups.Paths[freezer.Name()])
-	if err != nil {
-		multierror.Append(mErrs, fmt.Errorf("error getting pids: %v", err))
-
-		// Unfreeze the cgroup.
-		groups.Resources.Freezer = lconfigs.Thawed
-		freezer := cgroupFs.FreezerGroup{}
-		if err := freezer.Set(groups.Paths[freezer.Name()], groups); err != nil {
-			multierror.Append(mErrs, fmt.Errorf("failed to unfreeze cgroup: %v", err))
-			return mErrs.ErrorOrNil()
-		}
-	}
-
-	// Kill the processes in the cgroup
-	for _, pid := range pids {
-		proc, err := os.FindProcess(pid)
-		if err != nil {
-			multierror.Append(mErrs, fmt.Errorf("error finding process %v: %v", pid, err))
-			continue
-		}
-
-		procs = append(procs, proc)
-		if e := proc.Kill(); e != nil {
-			multierror.Append(mErrs, fmt.Errorf("error killing process %v: %v", pid, e))
-		}
-	}
-
-	// Unfreeze the cgroug so we can wait.
-	groups.Resources.Freezer = lconfigs.Thawed
-	if err := freezer.Set(groups.Paths[freezer.Name()], groups); err != nil {
-		multierror.Append(mErrs, fmt.Errorf("failed to unfreeze cgroup: %v", err))
-		return mErrs.ErrorOrNil()
-	}
-
-	// Wait on the killed processes to ensure they are cleaned up.
-	for _, proc := range procs {
-		// Don't capture the error because we expect this to fail for
-		// processes we didn't fork.
-		proc.Wait()
-	}
-
-	// Remove the cgroup.
-	if err := cgroups.RemovePaths(groups.Paths); err != nil {
-		multierror.Append(mErrs, fmt.Errorf("failed to delete the cgroup directories: %v", err))
-	}
 	return mErrs.ErrorOrNil()
 }
